Making an MVC Model - 1. Separate View and Model (and TimelineFactory class that is in essence a (producer of) model).  The View is the FXML's controller. 2. Pass model and TimelineFactory in setters to the View because the View is created by the FXMLoader.load() - following http://stackoverflow.com/questions/9717852/how-to-pass-object-created-in-fxml-controller1-to-controller2-of-inner-fxml-cont/10718683#10718683
Started change to using timelines as models to build the view.  Use a factory to get the timeline type - global, month, week, day.

Setting the event handler for the timeline menuItems in code with an aim to create a pattern of NOT using events on the FXML but just Ids.